+++ mr/src/main/java/org/elasticsearch/hadoop/serialization/handler/read/impl/AbortOnlyHandlerLoader.java
@@ -17,10 +17,12 @@
  * under the License.
  */
 
-package org.elasticsearch.hadoop.serialization.handler.read;
+package org.elasticsearch.hadoop.serialization.handler.read.impl;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import org.elasticsearch.hadoop.serialization.handler.read.DeserializationErrorHandler;
 
 /**
  * A handler loader that ignores the configured error handlers and only loads the abort handler.
